feat: add websocket functions, project completed.

+++ src/http/routes/vote-on-poll.ts
@@ -2,18 +2,20 @@ import { FastifyInstance } from "fastify"
 import {z} from "zod"
 import { randomUUID } from "crypto"
 import { prisma } from "../../lib/prisma"
+import { redis } from "../../lib/redis"
+import { voting } from "../../utils/Voting-pub-sub"
 
 const voteOnPoll = async (app: FastifyInstance) => {
   app.post('/polls/:pollId/votes', async (request, replay) => {
     const voteOnPollBody = z.object({
-      pollOptionsId: z.string().uuid()
+      pollOptionId: z.string().uuid()
     })
 
     const voteOnPollParams = z.object({
       pollId: z.string().uuid()
     })
 
-    const { pollOptionsId } = voteOnPollBody.parse(request.body)
+    const { pollOptionId } = voteOnPollBody.parse(request.body)
     const { pollId } = voteOnPollParams.parse(request.params)
 
     let { sessionId } = request.cookies
@@ -28,12 +30,20 @@ const voteOnPoll = async (app: FastifyInstance) => {
         }
       })
 
-      if (userPreviousVoteOnPoll) {
+      if (userPreviousVoteOnPoll && userPreviousVoteOnPoll.pollOptionId !== pollOptionId) {
         await prisma.vote.delete({
           where: {
             id: userPreviousVoteOnPoll.id
           }
         })
+
+      const votes = await redis.zincrby(pollId, -1, userPreviousVoteOnPoll.pollOptionId)
+
+      voting.publish(pollId, {
+        pollOptionId: userPreviousVoteOnPoll.pollOptionId,
+        votes: Number(votes),
+      })
+
       } else if (userPreviousVoteOnPoll) {
         return replay.status(400).send({message: "you already voted on this poll."})
       }
@@ -54,8 +64,15 @@ const voteOnPoll = async (app: FastifyInstance) => {
       data: {
         sessionId,
         pollId,
-        pollOptionsId
+        pollOptionId
       }
+    })
+
+    const votes = await redis.zincrby(pollId, 1, pollOptionId)
+
+    voting.publish(pollId, {
+      pollOptionId,
+      votes: Number(votes),
     })
 
     return replay.status(201).send()
